added a custom hook to postDetails
